api/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
httpx>=0.24.0
python-jose[cryptography]>=3.3.0
email-validator>=2.0.0
python-dotenv>=0.19.0
asyncpg>=0.27.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
psycopg2-binary==2.9.9
alembic==1.13.1 